Combine TypeChat with TypeChat.Schema (#91)
Moved all files from TypeChat.Schema to TypeChat. Updated namespaces
where needed and modified to the example projects to build with the new
structure. For the sake of simplicity, targeted TypeChat to .NET 7.0.
Future PR will migrate the project to standard 2.0.

Additionally added a new JsonTranslator constructor that allows for a
1-line constructor demonstrated in the examples. For simple apps, this
makes out-of-the-box usage much more intuitive.